post 7655 schedule line savings [pr]
